refactored, renamed enum and model files
