+++ src/app/shared/models/ad-response.model.ts
