+++ src/app/shared/models/enum/article-state.enum.ts
