+++ src/app/shared/models/enum/gender.enum.ts
@@ -1,5 +1,5 @@
 export enum Gender {
   MALE = 'Homme',
   FEMALE = 'Femme',
-  UNISEX = 'Unisex',
+  UNISEX = 'Unisexe',
 }
