+++ src/app/shared/models/enum/price-range.enum.ts
